pyproject.toml
[project]
name = "fastapi-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.116.1",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
]
